Add a publicly accessible landing page
So that we can health-check the management application during the deploy process, we need at least one page that is publicly accessible outside of CAS authentication. This opens management#index up outside of CAS, but only displays the actual dashboard if the user is authenticated. Otherwise, it tells the user they need to sign in.